media: qcom: camss: vfe: Avoid unnecessary RUP command
Today, we submit a RUP command (reg_update) after queuing a new frame
buffer using vfe_wm_update. However, vfe_wm_update writes the buffer
address to a FIFO register which does not require a subsequent RUP
command, as this register is not double-buffered like other config
or control shadow registers.

This results in the capture process generating twice the expected number
of interrupts, with one RUP done and one BUF done interrupt for each frame.

We can relax the interrupt pressure and VFE load by removing this extra
RUP command, which is only necessary when initializing or reconfiguring
the VFE.

Note: this has only been tested with vfe-340/qcm2290, the assumption is
that the other VFEs behave similarly.

Signed-off-by: Loic Poulain <loic.poulain@oss.qualcomm.com>