firmware: psci: Cold-reset as last resort warm-reset fallback
The PSCI SYSTEM_RESET2 function is optional and allows performing
a warm reset. If this function is not supported, we currently
substitute it with a cold reset using SYSTEM_RESET.

However, the platform may have other reboot handlers (with lower
priority) that do support warm reset. We should then give these
handlers a chance to execute the warm reset before resorting to
a cold reset.

This can be achieved by implementing a secondary reset handler
with lower priority.

Signed-off-by: Loic Poulain <loic.poulain@oss.qualcomm.com>